modify utils.ts
